+++ experimental/CHANGELOG.md
@@ -18,6 +18,7 @@ For notes on migrating to 2.x / 0.200.x see [the upgrade guide](doc/upgrade-to-2
 * feat(configuration): export interfaces required in other packages [#6462](https://github.com/open-telemetry/opentelemetry-js/pull/6462) @maryliag
 * feat(configuration): Zod-based codegen with reproducible generation and factory-layer defaults [#6498](https://github.com/open-telemetry/opentelemetry-js/pull/6498) @MikeGoldsmith
 * feat(configuration): set MeterProvider on sdk start [#6463](https://github.com/open-telemetry/opentelemetry-js/pull/6463) @maryliag
+* feat(configuration): export interfaces required in other packages [#6507](https://github.com/open-telemetry/opentelemetry-js/pull/6507) @maryliag
 
 ### :bug: Bug Fixes